+++ src/test/java/com/project/vegetable/repository/CustomerRepositoryTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Optional;
 
@@ -46,7 +47,7 @@ public class CustomerRepositoryTest {
         LocalDateTime orderDate = LocalDateTime.now();
 
         Order order = new Order();
-        order.setTotalAmount(20.00);
+        order.setTotalAmount(BigDecimal.valueOf(20.00));
         order.setCustomer(customer);
         order.setOrderDate(orderDate);
         orderRepository.save(order);
@@ -67,7 +68,7 @@ public class CustomerRepositoryTest {
         LocalDateTime orderDate = LocalDateTime.now();
 
         Order order = new Order();
-        order.setTotalAmount(20.00);
+        order.setTotalAmount(BigDecimal.valueOf(20.00));
         order.setCustomer(customer);
         order.setOrderDate(orderDate);
         orderRepository.save(order);
