+++ src/test/java/com/project/vegetable/repository/ProductRepositoryTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -25,7 +26,7 @@ public class ProductRepositoryTest {
 
         Product product = new Product();
         product.setName("Carrot");
-        product.setPrice(2.49f);
+        product.setPrice(BigDecimal.valueOf(2.49f));
         product.setStockQuantity(50);
         product.setCategory(category);
         return productRepository.saveAndFlush(product);
